// ui/src/generated/InputList.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AudioInputChannel } from "./AudioInputChannel";
import type { MidiPort } from "./MidiPort";

export type InputList = { audio: Array<AudioInputChannel>, midi: Array<MidiPort>, };