// ui/src/generated/TrackInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TrackId } from "./TrackId";

/**
 * Where a track takes its live input from (recording / monitoring).
 */
export type TrackInput = { "type": "None" } | { "type": "Audio", first: number, count: number, } | { "type": "Midi", port: string | null, channel: number | null, } | { "type": "Track", track: TrackId, };